updated hw3

+++ kernel/proc.h
@@ -104,7 +104,7 @@ struct proc {
   struct inode *cwd;           // Current directory
   char name[16];               // Process name (debugging)
 
-  uint64 priority;
-  uint64 readytime;
+  int priority;
+  int readytime;
   int cputime;
 };
